refactor(consumers): migrate Python branch/rankWord callers onto taxonomy.py authority
Collapse all Python consumers of trustMagnitude.computeBranch and
formatting.rank_word onto taxonomy.branchFor / taxonomy.rankWord:

  scripts/generateBadges.py       — import swap (branchFor, rankWord)
  scripts/generateOgCards.py      — import swap
  scripts/generateProfilePages.py — import swap
  scripts/migrate_taxonomy_v6.py  — import swap + drop genericSkillMap arg
  src/gaia_cli/promotion.py       — import swap + drop genericSkillMap arg

Shim the two deprecated functions to delegate rather than delete:
  trustMagnitude.computeBranch  — body replaced with branchFor() delegation
  formatting.rank_word          — body replaced with rankWord() delegation
  formatting.format_rank_label  — body replaced with rankWord() delegation

Shims are retained for backward-compat imports surviving until PR3b
(JS consumer sweep + hard grep guard). Zero new test failures.
Not-Basics-only gate: 23 suite/unique entries confirmed in named index.
