+++ scripts/generateBadges.py
@@ -42,8 +42,8 @@ sys.path.insert(0, str(REPO_ROOT / "scripts"))
 sys.path.insert(0, str(REPO_ROOT / "src"))
 from _atlas_helpers import named_slug  # noqa: E402
 from gaia_cli.redaction import REDACTED_HANDLE, is_redacted  # noqa: E402  single source of truth
-from gaia_cli.trustMagnitude import computeBranch as _compute_branch  # noqa: E402  Ygg-II read-time branch
-from gaia_cli.formatting import rank_word as _rank_word  # noqa: E402  branch-forked rank vocabulary
+from gaia_cli.taxonomy import branchFor as _compute_branch  # noqa: E402  Ygg-II branch authority
+from gaia_cli.taxonomy import rankWord as _rank_word  # noqa: E402  branch-forked rank vocabulary
 
 
 # Origin contributors render in GOLD (Yggdrasil II rubric E4: the deprecated
